ExecFromMain hook
