+++ src/commands/ScriptCommandRelay.h
@@ -31,9 +31,6 @@ public:
    static void StartScriptServer(tpRegScriptServerFunc scriptFn);
 };
 
-// The void * return is actually a Lisp LVAL and will be cast to such as needed.
-extern void * ExecForLisp( char * pIn );
-extern void * nyq_make_opaque_string( int size, unsigned char *src );
-extern void * nyq_reformat_aud_do_response(const wxString & Str);
+extern int ExecFromMain(wxString *pIn, wxString *pOut);
 
 #endif /* End of include guard: __SCRIPT_COMMAND_RELAY__ */
